fix golang template for purchase

+++ server/database/products.sql.go
@@ -236,7 +236,7 @@ func (q *Queries) GetBasicProduct(ctx context.Context, id string) (Product, erro
 }
 
 const getNotifyPurchaseDetails = `-- name: GetNotifyPurchaseDetails :one
-SELECT email, firstname, lastname, product_id, paid, chosen_date, title, (cost_baisa / 1000) as cost
+SELECT email, firstname, lastname, product_id, paid, chosen_date, title, (cost_baisa::FLOAT / 1000) as cost
 FROM purchases INNER JOIN users ON purchases.user_id = users.id INNER JOIN products ON purchases.product_id = products.id
 WHERE purchases.id = $1
 `
